+++ src/modules/project.nix
@@ -1,6 +1,11 @@
-{lib, ...}: let
+{
+  lib,
+  pkgs,
+  ...
+}: let
   l = lib // builtins;
   t = l.types;
+  nixpkgsRustLib = import "${pkgs.path}/pkgs/build-support/rust/lib" {inherit lib;};
 in {
   imports = [../options/drvConfig.nix];
   options = {
@@ -40,6 +45,29 @@ in {
       description = ''
         `profiles` option that will affect all packages in this project.
         For more information refer to `nci.crates.<name>.profiles` option.
+      '';
+    };
+    targets = l.mkOption {
+      type = t.attrsOf (t.submoduleWith {
+        modules = [./target.nix];
+      });
+      default = {
+        ${nixpkgsRustLib.toRustTarget pkgs.stdenv.hostPlatform}.default = true;
+      };
+      defaultText = ''
+        {
+          <host platform>.default = true;
+        }
+      '';
+      example = l.literalExpression ''
+        {
+          wasm32-unknown-unknown.profiles = ["release"];
+          x86_64-unknown-linux-gnu.default = true;
+        }
+      '';
+      description = ''
+        `targets` option that will affect all packages in this project.
+        For more information refer to `nci.crates.<name>.targets` option.
       '';
     };